requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
black==22.10.0
cached-property==1.5.2
ccxt==1.90.31
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cplex==22.1.0.0
cryptography==37.0.4
cytoolz==0.12.0
dataclassy==0.11.1
docplex==2.23.222
eip712==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.2
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b6
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
Flask==2.2.2
frozenlist==1.3.1
hexbytes==0.2.3
hypothesis==6.27.3
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.21.6
packaging==21.3
parsimonious==0.8.1
pathspec==0.10.1
Pillow==9.3.0
pinatapy-vourhey==0.1.8
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycares==4.2.1
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
pytz==2022.1
PyYAML==5.4.1
rdkit==2022.9.2
requests==2.28.1
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typed-ast==1.5.4
typing_extensions==4.3.0
uniswap-python==0.5.5
urllib3==1.26.12
varint==1.0.2
vvm==0.1.0
vyper==0.3.7
wcwidth==0.2.5
web3==5.31.1
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
